PV1.19 is defined as NM in the v2.2 spec
In v2.2 of the spec there's a PV1.19 is defined as a NM type but in
nHapi it's being mapped as the CM_PAT_ID type. I've created a patch to
fix this and attached it(not sure if it will get through the sf mail).
It does mention in the spec that PV1.19 is left as NM for backwards
compatibility but it is recommended to use the CK type instead. Maybe
that's why the type is different. My opinion is that if you're going to
match the spec then this should still be NM.