Fix issue where LocalGMTOffset always returns 0.
The `ZONE_OFFSET` flag is not supported in the CalendarManager, leading
to `DataTypeUtil.LocalGMTOffset` always returning 0.

To fix this we use the native .NET TimeZoneInfo libraries to provide the
correct Hour and Minute offsets.

Tests were added to handle "normal" time zones (PST, PDT), a timezone
with a non-zero minute offset (Myanmar @ 6h30m), and a timezone with a
14 hour offset (Line Islands).