Update CommonDT "Precision" setters to avoid memory leak
Similar to the CommonDT `Value` setter memory leak described in the
previous commit, the same issue applies to the "Precision" settings.

This commit fixes these leaks in the "Precision" setters.

Note that the `toHl7DTFormat` continues to leak memory, however it
doesn't really work as intended and is unlikely used in the wild (since
there's really no sane way to tie an arbritary Calendar from the
CalendarManager hashtable to any particular date).  The method is left
in place to avoid a contract breaking change.